==> notebooks/1.0-iac-data-splitting.rmd <==
---
title: "Data Splitting"
output: html_document
author: "Ivan A Chavez"
---

```{r}
# Loading packages and setting seed
library(caret)
library(tidyverse)
set.seed(100)
```

# Importing and assigning column names
```{r}
names <- c("SpMax_L", "J_Dz(e)", "nHM", "F01[N-N]", "F04[C-N]", "NssssC", "nCb-", "C%", "nCp", "nO", "F03[C-N]", "SdssC", "HyWi_B(m)", "LOC", "SM6_L", "F03[C-O]", "Me", "Mi", "nN-N", "nArNO2", "nCRX3", "SpPosA_B(p)", "nCIR", "B01[C-Br]", "B03[C-Cl]", "N-073", "SpMax_A", "Psi_i_1d", "B04[C-Br]", "SdO", "TI2_L", "nCrt", "C-026", "F02[C-N]", "nHDon", "SpMax_B(m)", "Psi_i_A", "nN", "SM6_B(m)", "nArCOOR", "nX", "experimental class")

biodeg <- read.csv("../data/biodeg.csv", header = FALSE, sep = ";", col.names = names)

#dropping extereme nCIR outlier row 1054
biodeg <- biodeg[-1054,]

```
# Plotting original frequency distributions of predictor variables
```{r}
biodeg_hist <- biodeg[,-42]
par(mfrow = c(7,6), mar=c(2,2,2,2))

for(i in names(biodeg_hist)){
  hist(biodeg_hist[[i]] ,main=i,xlab="x",label=TRUE,plot = TRUE)
}

```

```{r}
ggplot(biodeg, aes(experimental.class)) + geom_bar() + coord_flip()
```

# Implementing stratified sampling to maintain frequency distribution

```{r}

index_biodeg <- createDataPartition(biodeg$experimental.class, p=.80, list = FALSE)
biodeg_trainset <- biodeg[index_biodeg,]
biodeg_testset <- biodeg[-index_biodeg,]

response_train <- biodeg_trainset[,42]
response_test <- biodeg_testset[,42]

biodeg_train <- biodeg_trainset[,-42]
biodeg_test <- biodeg_testset[,-42]

```

# Verifying training set distributions

```{r}

par(mfrow = c(7,6), mar=c(2,2,2,2))

for(i in names(biodeg_train)){
  hist(biodeg_train[[i]] ,main=i,xlab="x",label=TRUE, plot = TRUE)
}

```

```{r}
ggplot(biodeg_trainset, aes(experimental.class)) + geom_bar() + coord_flip()
```

# Verifying test set distributions

```{r}

par(mfrow = c(7,6), mar=c(2,2,2,2))

for(i in names(biodeg_test)){
  hist(biodeg_test[[i]] ,main=i,xlab="x",label=TRUE, plot = TRUE)
}

```

```{r}
ggplot(biodeg_testset, aes(experimental.class)) + geom_bar() + coord_flip()
```

# Exporting train/test datasets for preprocessing

```{r}
write.csv(biodeg_train,"../data/biodeg_train.csv", row.names = FALSE)
write.csv(biodeg_test,"../data/biodeg_test.csv", row.names = FALSE)
write.csv(response_train,"../data/response_train.csv", row.names = FALSE)
write.csv(response_test,"../data/response_test.csv", row.names = FALSE)
```
