==> notebooks/1.0-sk-lda-svm-rf-model.rmd <==
---
title: "1.0-sk-lda"
author: "Sreeja K"
date: "`r Sys.Date()`"
output: word_document
---


```{r}
install.packages("mlbench")
library(caret)
library(tidyverse)
library(pROC)
library(mlbench)
library(randomForest)
set.seed(100)
```


```{r}
biodeg_train <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/biodeg_train.csv", header = TRUE, sep = ",")
biodeg_test <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/biodeg_test.csv", header = TRUE, sep = ",")

response_train <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/response_train.csv", header = TRUE, sep = ",")
response_test <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/response_test.csv", header = TRUE, sep = ",")

response_train <- as.factor(as.matrix(response_train))
response_test <- as.factor(as.matrix(response_test))

y_pred_test <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/test-results/y_pred.csv", header = TRUE, sep = ",")
y_pred_test <- lapply(y_pred_test, function(x) (as.factor(x)))
```


#Support_Vector_Machines

```{r}
ctrl <- trainControl(method = "cv", number = 5, summaryFunction = twoClassSummary, classProbs = TRUE,   savePredictions = TRUE)

#svm fit
#SVMFIT <- train(x = biodeg_train, y = response_train, method = "svmLinear", preProcess = c("center","scale"),  trControl = ctrl)

SVMFIT <- train(x = biodeg_train, y = response_train, method = "svmRadial", preProcess = c("center","scale"),  trControl = ctrl)

svm_pred<- data.frame(obs = response_test, SVM = predict(SVMFIT, biodeg_test))
svm_ROC = pROC::roc(response = response_train, predictor = svm_Pred[, 1])

svm_AUC = svm_ROC$auc[1]
svm = list(classifier = SVMFIT, predictions = svm_Pred, roc = svm_ROC, auc = svm_AUC)
SVMFIT
```



```{r}
Test_svm <- data.frame(obs = response_test,svm = predict(SVMFIT, biodeg_test))

```



#Linear_Discriminant_Analysis
```{r }
set.seed(100)
ctrl1 <- trainControl(method = "cv", number = 10, summaryFunction = twoClassSummary, classProbs = TRUE,   savePredictions = TRUE)
lda_fit = train(x=biodeg_train, y = response_train, method = "lda", preProc = c("center", "scale"), metric = "ROC", trControl = ctrl1)

lda_Pred  = predict(lda_fit,biodeg_train, type = "prob") 
lda_ROC = pROC::roc(response = response_train, predictor = lda_Pred[, 1])

lda_AUC = lda_ROC$auc[1]
lda = list(classifier = lda_fit, predictions = lda_Pred, roc = lda_ROC, auc = lda_AUC)
plot(lda_ROC, legacy.axes = TRUE)
```



```{r}
Test_lda <- data.frame(obs = response_test,lda = predict(lda_fit, biodeg_test))
```

#RandomForest

```{r }
set.seed(100)

rf_fit = randomForest(x=biodeg_train, y = response_train, preProc = c("center", "scale"), metric = "ROC", trControl = ctrl1)

rf_Pred  = predict(rf_fit,biodeg_test, type = "prob") 
rf_ROC = pROC::roc(response = response_train, predictor = rf_Pred[, 1])

rf_AUC = rf_ROC$auc[1]
rf = list(classifier = rf_fit, predictions = rf_Pred, roc = rf_ROC, auc = rf_AUC)
plot(rf_ROC, legacy.axes = TRUE)
```


```{r}
Test_rf <- data.frame(obs = response_test,rf = predict(rf_fit, biodeg_test))
```


```{r}
Models <- resamples(list(lda = lda_fit,
                         rf = rf_fit, 
                         svm = SVMFIT))
Summ_Mod <- summary(Models)
Summ_Mod

```

```{r}
confusionMatrix(Test_lda$lda, Test_lda$obs, positive = "RB")
confusionMatrix(Test_rf$rf, Test_rf$obs, positive = "RB")
confusionMatrix(Test_svm$svm, Test_svm$obs, positive = "RB")
```

```{r}
#writing the output
write.csv(y_pred_test,"C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/model-output/y_pred_sr.csv", row.names = FALSE)
```
