==> notebooks/1.0-sk-svm.rmd <==
---
title: "1.0-sk-svm"
author: "Sreeja K"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
install.packages("mlbench")
library(caret)
library(tidyverse)
library(pROC)
library(mlbench)
set.seed(100)
```


```{r}
biodeg_train <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/biodeg_train.csv", header = TRUE, sep = ",")
biodeg_test <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/biodeg_test.csv", header = TRUE, sep = ",")

response_train <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/response_train.csv", header = TRUE, sep = ",")
response_test <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/response_test.csv", header = TRUE, sep = ",")

response_train <- as.factor(as.matrix(response_train))
response_test <- as.factor(as.matrix(response_test))

y_pred_test <- read.csv("C:/Users/kurap/OneDrive/Documents/GitHub/predictive-modeling/data/test-results/y_pred.csv", header = TRUE, sep = ",")
y_pred_test <- lapply(y_pred_test, function(x) (as.factor(x)))
```



```{r}
ctrl <- trainControl(method = "repeatedcv", repeats = 5, summaryFunction = twoClassSummary, classProbs = TRUE, savePredictions = TRUE)

#svm fit
SVMFIT <- train(x = biodeg_train, y = response_train, method = "svmLinear", preProcess = c("center","scale"),  trControl = ctrl)
SVMFIT
```


```{r}
#SVM fit using cost
SVMFIT1 <- train(x = biodeg_train, y = response_train, method = "svmLinear", tuneGrid = expand.grid(C = seq(0, 2, length = 20)), trControl = ctrl)
SVMFIT1

plot(SVMFIT1)

#finds the best tuning parameter
SVMFIT1$bestTune
```

```{r}
SVM1_OP<-as_tibble(SVMFIT1$results[which.min(SVMFIT1$results[,2]),])
SVM1_OP
```

```{r}
#SVM Radial using non-linear kernel
SVMFIT2 <- train(x = biodeg_train, y = response_train, method = "svmRadial",  trControl = ctrl, preProcess = c("center","scale"), tuneLength = 10)
SVMFIT2

SVMFIT2$bestTune
```

```{r}
SVM2_OP<-as_tibble(SVMFIT2$results[which.min(SVMFIT2$results[,2]),])
SVM2_OP
```

```{r}
#SVM poly
SVMFIT3 <- train(x = biodeg_train, y = response_train, method = "svmPoly", trControl = ctrl, preProcess = c("center","scale"), tuneLength = 10)
SVMFIT3

SVMFIT3$bestTune
```

```{r}
SVM3_OP<-as_tibble(SVMFIT3$results[which.min(SVMFIT3$results[,2]),])
SVM3_OP
```

```{r}
df<-tibble(Model=c('SVM_Linear','SVM_Linear using cost','SVM_Radial','SVM_Poly'), sensitivity=c(SVMFIT$results[2][[1]],SVM1_OP$Sens,SVM2_OP$Sens,SVM3_OP$Sens))
df %>% arrange(sensitivity)
```
