==> notebooks/2.0-iac-nsc.rmd <==
---
title: "Nearest Shrunken Centroid Model"
output: html_document
author: "Ivan A Chavez"
---
```{r}
# Loading packages and setting seed
library(caret)
set.seed(100)
```
```{r}
# Loading transformed data to perform modeling
biodeg_train <- read.csv("../data/biodeg_train.csv", header = TRUE, sep = ",")
biodeg_test <- read.csv("../data/biodeg_test.csv", header = TRUE, sep = ",")
response_train <- read.csv("../data/response_train.csv", header = TRUE, sep = ",")
response_test <- read.csv("../data/response_test.csv", header = TRUE, sep = ",")

response_train <- as.factor(as.matrix(response_train))
response_test <- as.factor(as.matrix(response_test))

```


```{r}

ctrl <- trainControl(method = "cv",
                       number = 5,
                        summaryFunction = twoClassSummary,
                        classProbs = TRUE,
                        savePredictions = TRUE)

ctrl_10 <- trainControl(method = "cv",
                     summaryFunction = twoClassSummary,
                     classProbs = TRUE,
                     savePredictions = TRUE)

nscFit_noPre <- train(x = biodeg_train,
                y = response_train,
                method = "pam",
                tuneGrid = data.frame(threshold = seq(0, 25, length = 30)),
                metric = "ROC",
                trControl = ctrl)

nscFit_cs <- train(x = biodeg_train,
                y = response_train,
                method = "pam",
                tuneGrid = data.frame(threshold = seq(0, 25, length = 30)),
                preProcess = c('center','scale'),
                metric = "ROC",
                trControl = ctrl)

nscFit_10_noPre <- train(x = biodeg_train,
                      y = response_train,
                      method = "pam",
                      tuneGrid = data.frame(threshold = seq(0, 25, length = 30)),
                      metric = "ROC",
                      trControl = ctrl_10)

nscFit_10_cs <- train(x = biodeg_train,
                   y = response_train,
                   method = "pam",
                   tuneGrid = data.frame(threshold = seq(0, 25, length = 30)),
                   preProcess = c('center','scale'),
                   metric = "ROC",
                   trControl = ctrl_10)

nscFit_10_cs

#nscFit_5

```


```{r}

test_preds <- data.frame(obs = response_test,
                         M8_NSC_no_pre = predict(nscFit_noPre, biodeg_test))

test_preds$M8_NSC_CS <- predict(nscFit_cs, biodeg_test)
test_preds$M8_NSC_10_no_pre <- predict(nscFit_10_noPre, biodeg_test)
test_preds$M8_10_NSC_CS <- predict(nscFit_10_cs, biodeg_test)



models_compare <- resamples(list(M8_NSC_no_pre = nscFit_noPre, M8_NSC_CS = nscFit_cs))

models_compare_summ <- summary(models_compare)
models_compare_summ

models_compare_10 <- resamples(list(M8_NSC_10_no_pre = nscFit_10_noPre, M8_10_NSC_CS = nscFit_10_cs))

models_compare_summ_10 <- summary(models_compare_10)
models_compare_summ_10


```

```{r}
confusionMatrix(test_preds$M8_10_NSC_CS, test_preds$obs, positive = "RB")
```

```{r}

write.csv(data.frame(models_compare_summ$models, models_compare_summ$statistics),
          "../data/resampling-results/resamp_results_nsc_5.csv", row.names = FALSE)

write.csv(data.frame(models_compare_summ_10$models, models_compare_summ_10$statistics),
          "../data/resampling-results/resamp_results_nsc_10.csv", row.names = FALSE)

write.csv(test_preds[,-1],"../data/model-output/y_pred_nsc.csv", row.names = FALSE)
```
